Refactor SetPassword to use base view model and subclasses
- Rename setPassword.js -> passwordForms.js
- No longer need to pass 'claim'/'reset'
- Remove unnecessary constructor arguments and methods
