+++ website/static/js/pages/login-page.js
@@ -4,12 +4,8 @@
 'use strict';
 var $ = require('jquery');
 
-var SignUp = require('js/setPassword');
-var LogInForm = require('js/signIn');
+var passwordForms = require('js/passwordForms');
 var InstitutionSignIn = require('js/institutionSignIn');
-
-var registerUrl = window.contextVars.registerUrl;
-
 
 var activateToggleBox = function () {
     var el = $(this);
@@ -25,8 +21,7 @@ $(document).ready(function() {
     if (campaign === 'institution'){
         new InstitutionSignIn('#inst');
     } else {
-        new LogInForm.SignIn('#logInForm');
-        new SignUp('#signUpScope', 'signup', registerUrl, campaign);
+        new passwordForms.SignUp('#signUpScope');
     }
 });
